fixed small bug + added capability to estimate CIs
small bug: the custom sign_() function was encountering some issues when numpy was already imported in the working environment, as some booleans were interpreted as numpy booleans and could not subtract them ("numpy boolean subtract, the `-` operator, is not supported, use the bitwise_xor, the `^` operator, or the logical_xor function instead."), when with regular boolean you can (True-False=1 etc.). Anyway, I did this function to avoid imports but as I already was forced to import to have the probability functions I just imported the numpy.sign, and problem fixed.

the functions now also return the +/- value of the 95% confidence intervals of slope and intercept.